microservice/requirements.txt
Flask
sentence-transformers
pymongo
numpy==1.24.3
python-dotenv
torch